[TASK] Fix phpstan checkFunctionArgumentTypes errors in several extensions
This patch fixes incompatible type usage in function arguments
and is preparatory work for introducing native type hints and
strict mode in all core files.

Releases: master, 10.4
Resolves: #92161
Resolves: #92164
Resolves: #92175
Resolves: #92176
Resolves: #92177
Resolves: #92178
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65868
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>

+++ typo3/sysext/dashboard/Classes/Widgets/Provider/SysLogErrorsDataProvider.php
@@ -101,9 +101,9 @@ class SysLogErrorsDataProvider implements ChartDataProviderInterface
         $format = $GLOBALS['TYPO3_CONF_VARS']['SYS']['ddmmyy'] ?: 'Y-m-d';
 
         for ($daysBefore = $this->days; $daysBefore >= 0; $daysBefore--) {
-            $this->labels[] = date($format, strtotime('-' . $daysBefore . ' day'));
-            $startPeriod = strtotime('-' . $daysBefore . ' day 0:00:00');
-            $endPeriod =  strtotime('-' . $daysBefore . ' day 23:59:59');
+            $this->labels[] = date($format, (int)strtotime('-' . $daysBefore . ' day'));
+            $startPeriod = (int)strtotime('-' . $daysBefore . ' day 0:00:00');
+            $endPeriod =  (int)strtotime('-' . $daysBefore . ' day 23:59:59');
 
             $this->data[] = $this->getNumberOfErrorsInPeriod($startPeriod, $endPeriod);
         }
